+++ src/page/messagePage/MyListPage.js
@@ -97,7 +97,9 @@ const MyListPage = () => {
                             className="icon"
                             onClick={(e) => {
                               e.stopPropagation();
-                              deleteMessage(message.id);
+                              if (window.confirm("쪽지를 삭제하시겠습니까?")) {
+                                deleteMessage(message.id);
+                              }
                             }}
                           />
                         </span>
